Add files via upload
the employee database schema is a company database under FreecodeCamp's SQL tutorial guide, with inserted values created using MySQL in the popsql interface and exported to Excel. It contains various datatypes, primary and linked keys. 
the customer database is a simple database I created usingMS access to create the table, build a form and create queries.